Refactored the update unit tests.
